requirements_freeze.txt
anyio==4.12.0
certifi==2025.11.12
click==8.3.1
contourpy==1.3.2
cycler==0.12.1
exceptiongroup==1.3.1
filelock==3.20.0
fonttools==4.61.1
fsspec==2025.12.0
grad-cam==1.5.5
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.3
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
lightning-utilities==0.15.2
MarkupSafe==2.1.5
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
opencv-python==4.12.0.88
openslide-python==1.4.3
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
sympy==1.13.1
threadpoolctl==3.6.0
timm==1.0.22
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchmetrics==1.8.2
torchvision==0.20.1+cu121
tqdm==4.67.1
triton==3.1.0
ttach==0.0.3
typer-slim==0.21.0
typing_extensions==4.15.0
tzdata==2025.3
